% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{getAudioCollections}
\alias{getAudioCollections}
\title{Get the details of audio collections}
\source{
\dQuote{\code{/audio/collections/\{id\}}}
}
\usage{
getAudioCollections(id, ...)
}
\arguments{
\item{id}{character. Collection ID. (required)}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Get the details of audio collections
}
\examples{
\dontrun{
getAudioCollections(id = "48433107")
}
}
