% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{getAudioCollectionsContent}
\alias{getAudioCollectionsContent}
\title{Get the contents of audio collections}
\source{
\dQuote{\code{/audio/collections/\{id\}/items}}
}
\usage{
getAudioCollectionsContent(id, ...)
}
\arguments{
\item{id}{character. Collection ID. (required)}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Get the contents of audio collections
}
\examples{
\dontrun{
getAudioCollectionsContent(id = "48433113")
}
}
