% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{getAudioDetails}
\alias{getAudioDetails}
\title{Get the details of image collections}
\source{
\dQuote{\code{/audio/\{id\}}}
}
\usage{
getAudioDetails(id, ...)
}
\arguments{
\item{id}{character. Audio track ID. (required)}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Get the details of image collections
}
\examples{
\dontrun{
getAudioDetails(id = "442583")
}
}
