% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributors.R
\name{getContributorCollectionsDetails}
\alias{getContributorCollectionsDetails}
\title{Get details about contributors' collections}
\source{
\dQuote{\code{/contributors/\{contributor_id\}/collections\{id\}}}
}
\usage{
getContributorCollectionsDetails(contributor_id, id, ...)
}
\arguments{
\item{contributor_id}{character. Contributor ID. (required)}

\item{id}{character. Collection ID. (required)}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Get details about contributors' collections
}
\examples{
\dontrun{
getContributorCollectionsDetails(contributor_id = "800506", id = "1991678")
}
}
