% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributors.R
\name{getContributorCollectionsDetailsItems}
\alias{getContributorCollectionsDetailsItems}
\title{Get the items in contributors' collections}
\source{
\dQuote{\code{/contributors/\{contributor_id\}/collections\{id\}/items}}
}
\usage{
getContributorCollectionsDetailsItems(contributor_id, id, ...)
}
\arguments{
\item{contributor_id}{character. Contributor ID. (required)}

\item{id}{character. Collection ID. (required)}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Get the items in contributors' collections
}
\examples{
\dontrun{
getContributorCollectionsDetailsItems(contributor_id = "800506", id = "1991678", sort = "newest")
}
}
