% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributors.R
\name{getContributorDetails}
\alias{getContributorDetails}
\title{Get details about contributors}
\source{
\dQuote{\code{/contributors}}
}
\usage{
getContributorDetails(id, ...)
}
\arguments{
\item{id}{character. Contributor ID. (required)}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Get details about contributors
}
\examples{
\dontrun{
getContributorDetails(id = c("800506", "1653538"))
}
}
