% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editorial.R
\name{getEditorialDetails}
\alias{getEditorialDetails}
\title{Get details about editorial content}
\source{
\dQuote{\code{/editorial/\{id\}}}
}
\usage{
getEditorialDetails(id, ...)
}
\arguments{
\item{id}{character. Editorial ID.}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Get details about editorial content
}
\examples{
\dontrun{
getEditorialDetails(id = "9926131a")
}
}
