% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editorial.R
\name{getEditorialLivefeed}
\alias{getEditorialLivefeed}
\title{Get editorial livefeed}
\source{
\dQuote{\code{/editorial/livefeeds/\{id\}}}
}
\usage{
getEditorialLivefeed(id, ...)
}
\arguments{
\item{id}{character. Editorial livefeed ID (in URI encoded string form).}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Get editorial livefeed
}
\examples{
\dontrun{
getEditorialLivefeed(
id = "2018\%2F10\%2F15\%2FWomen\%20of\%20the\%20Year\%20Lunch\%20\%26\%20Awards\%2C\%20London", country = "USA"
)
}
}
