% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editorial.R
\name{getEditorialLivefeedList}
\alias{getEditorialLivefeedList}
\title{Get editorial livefeed list}
\source{
\dQuote{\code{/editorial/livefeeds}}
}
\usage{
getEditorialLivefeedList(...)
}
\arguments{
\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Get editorial livefeed list
}
\examples{
\dontrun{
getEditorialLivefeedList(country = "TR")
}
}
