% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{getFeaturedCollections}
\alias{getFeaturedCollections}
\title{Get the details of featured collections}
\source{
\dQuote{\code{/images/collections/featured/\{id\}}}
}
\usage{
getFeaturedCollections(id, ...)
}
\arguments{
\item{id}{character. Collection ID. (required)}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Giving more detailed information about a featured collection such as its
cover image, creation and most recent update time.
}
\examples{
\dontrun{
getFeaturedCollections(id = "136351027")
}
}
