% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{getFeaturedCollectionsItems}
\alias{getFeaturedCollectionsItems}
\title{Get the contents of featured collections}
\source{
\dQuote{\code{/images/collections/featured/\{id\}/items}}
}
\usage{
getFeaturedCollectionsItems(id, ...)
}
\arguments{
\item{id}{character. Collection ID. (required)}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Listing the image IDs and date when added to a featured collection.
}
\examples{
\dontrun{
getFeaturedCollectionsItems(id = "136351027")
}
}
