% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{getImageCollections}
\alias{getImageCollections}
\title{Get the details of image collections}
\source{
\dQuote{\code{/images/collections/\{id\}}}
}
\usage{
getImageCollections(id, ...)
}
\arguments{
\item{id}{character. Collection ID. (required)}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Giving more detailed information about a collection such as its cover image,
creation and most recent update time.
}
\examples{
\dontrun{
# enter an id from the collections you own:
getImageCollections(id = "139149928")
}
}
