% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{getImageCollectionsContent}
\alias{getImageCollectionsContent}
\title{Get the contents of image collections}
\source{
\dQuote{\code{/images/collections/\{id\}/items}}
}
\usage{
getImageCollectionsContent(id, ...)
}
\arguments{
\item{id}{character. Collection ID. (required)}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Get the contents of image collections
}
\examples{
\dontrun{
# enter an id from the collections you own:
getImageCollectionsContent(id = "139149928")
}
}
