% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{getImageCollectionsDetails}
\alias{getImageCollectionsDetails}
\title{Get the details of image collections}
\source{
\dQuote{\code{/images/collections/\{id\}/items}}
}
\usage{
getImageCollectionsDetails(id, ...)
}
\arguments{
\item{id}{character. Collection ID. (required)}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Listing the image IDs and date when added to a collection.
}
\examples{
\dontrun{
getImageCollectionsDetails(id = "139149928")
}
}
