% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{getImageDetails}
\alias{getImageDetails}
\title{Get details about images}
\source{
\code{\dQuote{images/\{id\}}}
}
\usage{
getImageDetails(id, ...)
}
\arguments{
\item{id}{character. Image ID. (required)}
}
\description{
Showing information about an image, including a URL to a preview image and
the sizes that it is available in.
}
\examples{
\dontrun{
getImageDetails(id = "620154782")
}
}
