% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{getUserDetails}
\alias{getUserDetails}
\title{Get user details}
\source{
\dQuote{\code{/user}}
}
\usage{
getUserDetails(...)
}
\arguments{
\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Get user details
}
\examples{
\dontrun{
getUserDetails()
}
}
