% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\name{getVideoCollections}
\alias{getVideoCollections}
\title{Get the details of video collections}
\source{
\dQuote{\code{/videos/collections/\{id\}}}
}
\usage{
getVideoCollections(id, ...)
}
\arguments{
\item{id}{character. Collection ID. (required)}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Get the details of video collections
}
\examples{
\dontrun{
# enter an id from the collections you own:
getVideoCollections(id = "17555176")
}
}
