% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\name{getVideoCollectionsContent}
\alias{getVideoCollectionsContent}
\title{Get the contents of video collections}
\source{
\dQuote{\code{/videos/collections/\{id\}/items}}
}
\usage{
getVideoCollectionsContent(id, ...)
}
\arguments{
\item{id}{character. Collection ID. (required)}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Get the contents of video collections
}
\examples{
\dontrun{
# enter an id from the collections you own:
getVideoCollectionsContent(id = "139149928")
}
}
