% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{listAudioCollections}
\alias{listAudioCollections}
\title{List audio collections}
\source{
\dQuote{\code{/audio/collections}}
}
\usage{
listAudioCollections(...)
}
\arguments{
\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
List audio collections
}
\examples{
\dontrun{
listAudioCollections()
}
}
