% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{listAudioLicenses}
\alias{listAudioLicenses}
\title{List audio licenses}
\source{
\dQuote{\code{/audio/licenses}}
}
\usage{
listAudioLicenses(...)
}
\arguments{
\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
List audio licenses
}
\examples{
\dontrun{
listAudioLicenses()
listAudioLicenses(audio_id = 442583)
}
}
