% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{listAudioTracks}
\alias{listAudioTracks}
\title{List audio tracks}
\source{
\dQuote{\code{/audio}}
}
\usage{
listAudioTracks(id, ...)
}
\arguments{
\item{id}{character. One or more Audio IDs.}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
List audio tracks
}
\examples{
\dontrun{
listAudioTracks(id = "442583")
listAudioTracks(id = c("442583", "434750"))
}
}
