% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributors.R
\name{listContributorCollections}
\alias{listContributorCollections}
\title{List contributors' collections}
\source{
\dQuote{\code{/contributors/\{contributor_id\}/collections}}
}
\usage{
listContributorCollections(contributor_id, ...)
}
\arguments{
\item{contributor_id}{character. Contributor ID. (required)}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
List contributors' collections
}
\examples{
\dontrun{
listContributorCollections(contributor_id = "800506", sort = "newest")
}
}
