% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{listFeaturedCollections}
\alias{listFeaturedCollections}
\title{List featured collections}
\source{
\dQuote{\code{/images/collections/featured/}}
}
\usage{
listFeaturedCollections(...)
}
\arguments{
\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Listing featured collections based on their types.
}
\examples{
\dontrun{
listFeaturedCollections()
listFeaturedCollections(type = "photo")
}
}
