% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{listImageCategories}
\alias{listImageCategories}
\title{List image categories}
\source{
\code{\dQuote{images/categories}}
}
\usage{
listImageCategories()
}
\description{
Listing the categories (Shutterstock-assigned genres) that images can belong
to. \emph{This call takes no additional parameters.}
}
\examples{
\dontrun{
listImageCategories()
}
}
