% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{listImageCollections}
\alias{listImageCollections}
\title{List image collections}
\source{
\dQuote{\code{/images/collections/}}
}
\usage{
listImageCollections(...)
}
\arguments{
\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
List image collections
}
\examples{
\dontrun{
listImageCollections()
}
}
