% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{listLicenses}
\alias{listLicenses}
\title{List licenses}
\source{
\dQuote{\code{/images/licenses/}}
}
\usage{
listLicenses(...)
}
\arguments{
\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Listing existing licenses.
}
\examples{
\dontrun{
listLicenses(image_id = "620154782")
}
}
