% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{listRecommendedImages}
\alias{listRecommendedImages}
\title{List recommended images}
\source{
\dQuote{\code{/images/recommendations}}
}
\usage{
listRecommendedImages(id, ...)
}
\arguments{
\item{id}{character. One or more Image IDs.}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Returning images that customers put in the same collection as the specified
image IDs.
}
\examples{
\dontrun{
listRecommendedImages(id = "620154782")
listRecommendedImages(id = c("620154782", "620154770"), max_items = 5)
}
}
