% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\name{listSimilarVideos}
\alias{listSimilarVideos}
\title{List similar videos}
\source{
\dQuote{\code{/videos/\{id\}/similar}}
}
\usage{
listSimilarVideos(id, ...)
}
\arguments{
\item{id}{character. Video ID. (required)}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
List similar videos
}
\examples{
\dontrun{
listSimilarVideos(id = "2140697")
}
}
