% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{listUserSubscriptions}
\alias{listUserSubscriptions}
\title{List user subscriptions}
\source{
\dQuote{\code{/user/subscriptions}}
}
\usage{
listUserSubscriptions(...)
}
\arguments{
\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
List user subscriptions
}
\examples{
\dontrun{
listUserSubscriptions()
}
}
