% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\name{listVideoCategories}
\alias{listVideoCategories}
\title{List video categories}
\source{
\dQuote{\code{/videos/categories}}
}
\usage{
listVideoCategories(...)
}
\arguments{
\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
List video categories
}
\examples{
\dontrun{
listVideoCategories()
}
}
