% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\name{listVideoCollections}
\alias{listVideoCollections}
\title{List video collections}
\source{
\dQuote{\code{/videos/collections/}}
}
\usage{
listVideoCollections(...)
}
\arguments{
\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
List video collections
}
\examples{
\dontrun{
listVideoCollections()
}
}
