% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\name{listVideoLicenses}
\alias{listVideoLicenses}
\title{List video licenses}
\source{
\dQuote{\code{/videos/licenses}}
}
\usage{
listVideoLicenses(...)
}
\arguments{
\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
List video licenses
}
\examples{
\dontrun{
listVideoLicenses(video_id = "2140697")
}
}
