% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\name{listVideos}
\alias{listVideos}
\title{List videos}
\source{
\dQuote{\code{/videos}}
}
\usage{
listVideos(id, ...)
}
\arguments{
\item{id}{character. One or more Video IDs.}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
List videos
}
\examples{
\dontrun{
listVideos(id = c("18002566", "17139196"))
}
}
