% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{searchAudio}
\alias{searchAudio}
\title{Search tracks}
\source{
\dQuote{\code{/audio/search}}
}
\usage{
searchAudio(...)
}
\arguments{
\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Search tracks
}
\examples{
\dontrun{
searchAudio(query = "bluegrass", duration_from = "60", moods = "uplifting")
}
}
