% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editorial.R
\name{searchEditorial}
\alias{searchEditorial}
\title{Search editorial content}
\source{
\dQuote{\code{/editorial/search}}
}
\usage{
searchEditorial(...)
}
\arguments{
\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Search editorial content
}
\examples{
\dontrun{
searchEditorial(query = "history", country = "TR")
searchEditorial(query = "windmills", country = "NL", sort = "newest", date_start = "2018-11-01")
}
}
