% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{searchImages}
\alias{searchImages}
\title{Search images}
\source{
\code{\dQuote{images/search}}
}
\usage{
searchImages(...)
}
\arguments{
\item{...}{arguments to be passed by the parameters in the source endpoint.}
}
\description{
Searching for images.
}
\examples{
\dontrun{
searchImages(query = "paris")

today <- as.character(Sys.Date())
searchImages(query = "farmer", sort = "popular", added_date = today)
}
}
