% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videos.R
\name{searchVideos}
\alias{searchVideos}
\title{Search videos}
\source{
\dQuote{\code{/videos/search}}
}
\usage{
searchVideos(...)
}
\arguments{
\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Search videos
}
\examples{
\dontrun{
searchVideos(query = "hot air balloon", duration_from = "30", sort = "popular")
}
}
