% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{sstk_auth}
\alias{sstk_auth}
\title{Authorize Shutterstock}
\usage{
sstk_auth(scopes = NULL)
}
\arguments{
\item{scopes}{A list of scopes defined in the Shutterstock API.
Default value is set to \code{NULL} which Shutterstock grants the
\code{user.view} scope by default when no additional scopes have been
provided.}
}
\description{
Authenticate and store user credentials to authorize requests for the
Shutterstock API. After a successful authentication, an OAuth token will be
cached inside the \code{.httr-oauth}, which is a file placed in the current
working directory. This workflow is mainly followed by the \code{httr}
package.
}
\details{
See the OAuth scope list
\url{https://api-reference.shutterstock.com/#authentication-oauth-scopes-h2}
}
\examples{
\dontrun{
sstk_auth(scopes = c("collections.view", "licenses.view"))
}
}
