% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{sstk_get_keys}
\alias{sstk_get_keys}
\title{Read environment variables}
\usage{
sstk_get_keys()
}
\description{
Reading environment variables either specified in the \code{.Renviron} file,
which is a safe place to locally store and retrieve API keys, or directly
from the environment variables set via \code{\link{sstk_set_keys}}.
}
