% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{sstk_keys}
\alias{sstk_keys}
\title{Read environment variables}
\usage{
sstk_keys()
}
\description{
Reads environment variables from the \code{.Renviron} file, which is a safe
place to locally store and retrieve API keys.
}
