% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{sstk_set_keys}
\alias{sstk_set_keys}
\title{Set environment variables}
\usage{
sstk_set_keys(id, secret, callback)
}
\arguments{
\item{id}{character. The client id from your application.}

\item{secret}{character. The secret id from your application.}

\item{callback}{character. The callback (or redirect) URI specified in your
application.}
}
\description{
Set environment variables
}
