% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{getImageDetails}
\alias{getImageDetails}
\title{Get details about images}
\source{
\dQuote{\code{/images/\{id\}}}
}
\usage{
getImageDetails(id, ...)
}
\arguments{
\item{id}{character. Image ID. (required)}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Showing information about an image, including its URL and preview image and
the sizes that it is available in.
}
\examples{
\dontrun{
getImageDetails(id = "620154782")
}
}
