% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{listImageCategories}
\alias{listImageCategories}
\title{List image categories}
\source{
\dQuote{\code{/images/categories}}
}
\usage{
listImageCategories(...)
}
\arguments{
\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Listing the categories (Shutterstock-assigned genres) that images can belong
to.
}
\examples{
\dontrun{
listImageCategories()
}
}
