% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{listSimilarImages}
\alias{listSimilarImages}
\title{List similar images}
\source{
\dQuote{\code{/images/\{id\}/similar}}
}
\usage{
listSimilarImages(id, ...)
}
\arguments{
\item{id}{character. Image ID. (required)}

\item{...}{arguments to be passed by the source endpoint parameters.}
}
\description{
Returning the images that are visually similar to an image that you specify.
}
\examples{
\dontrun{
listSimilarImages(id = "620154782")
}
}
