% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shutterstock.R
\docType{package}
\name{shutterstock-package}
\alias{shutterstock}
\alias{shutterstock-package}
\title{Shutterstock API client}
\description{
The endpoint descriptions are retrieved from Shutterstock API Reference
documentation.
}
\section{Package options}{

\describe{
\item{\code{sstk.api.root.url}}{The base URL for retrieving the API
endpoints. Default: \code{https://api.shutterstock.com/}}
}
}

\references{
Shutterstock API Reference:
  \url{https://api-reference.shutterstock.com/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/strboul/shutterstock-r}
  \item Report bugs at \url{https://github.com/strboul/shutterstock-r/issues}
}

}
\author{
\strong{Maintainer}: Metin Yazici \email{stradivariusboul@gmail.com}

}
\keyword{internal}
