% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{sstk_auth}
\alias{sstk_auth}
\title{Authorize Shutterstock}
\usage{
sstk_auth()
}
\description{
Authenticate and store user credentials to authorize requests for the
Shutterstock API. After a successful authentication, an OAuth token will be
cached inside the \code{.httr-oauth}, which is a file placed in the current
working directory. This workflow is mainly followed by the \code{httr}
package.
}
